requirement.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.3.0
Django==5.1
django-cors-headers==4.4.0
django-heroku==0.3.1
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.8
inflection==0.5.1
packaging==24.1
psycopg2==2.9.10
PuLP==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.8.2
